+++ WeAr/index.js
@@ -4,7 +4,7 @@ import {AppRegistry} from 'react-native';
 import App from './App';
 import {name as appName} from './app.json';
 import { Provider } from 'react-redux';
-import configureStore from './js/state-management/store';
+import configureStore from './js/store';
 
 const store =  configureStore();
