+++ WeAr/js/actions/types.js
