+++ WeAr/js/reducers/filterReducer.js
